pyproject.toml
[project]
name = "python-cluedo"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=4.1.0,<5",
    "numpy>=2.4.4,<3",
    "pandas>=3.0.2,<4",
    "plotly>=6.6.0,<7",
    "pydantic-settings>=2.13.1",
    "python-sat[aiger,pblib]>=1.9.dev2,<2",
    "python-socketio>=5.16.1,<6",
    "ruff>=0.15.9",
    "uvicorn>=0.46.0,<0.47",
]

[dependency-groups]
dev = ["pytest>=9.0.3,<10"]
